Don't install Drush or Composer if they already exist.

+++ vdd/chef/cookbooks/core/vdd/recipes/composer.rb
@@ -4,4 +4,5 @@ bash "install-composer" do
   curl -sS https://getcomposer.org/installer | php
   mv composer.phar /usr/local/bin/composer
   EOH
+  not_if { ::File.exists?("/usr/local/bin/composer") }
 end
